% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get_base_url}
\alias{cache_get_base_url}
\title{cache_get_base_url}
\usage{
cache_get_base_url(url)
}
\arguments{
\item{url}{}
}
\value{
the current_base_url as a string
}
\description{
returns the current base_url
}
