% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get_response}
\alias{cache_get_response}
\title{cache_get_response}
\usage{
cache_get_response(key)
}
\arguments{
\item{key}{key in the response store}
}
\value{
the
}
\description{
cache_get_response
}
