% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_get_token}
\alias{cache_get_token}
\title{cache_get_token}
\usage{
cache_get_token()
}
\value{
the jwt token used for request headers
}
\description{
Retrieves the current api token for use in constructing request headers for API
endpoints
}
