% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set_base_url}
\alias{cache_set_base_url}
\title{cache_set_base_url}
\usage{
cache_set_base_url(url)
}
\arguments{
\item{url}{the base url for the project}
}
\value{
the set url string or NULL
}
\description{
Set the base url. Will override configuration set by the cache_config_base_url setting
}
