% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set_response}
\alias{cache_set_response}
\title{cache_set_response}
\usage{
cache_set_response(key, data)
}
\arguments{
\item{key}{the key to store the data object}

\item{data}{the data object}
}
\description{
Used to store the results of requests. This could be json objects from the response
body or dataframes/vectors and other R objects for use by apps.
}
