% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_set_token}
\alias{cache_set_token}
\title{cache_set_token}
\usage{
cache_set_token(token)
}
\arguments{
\item{token}{the jwt token sent back from the auth}
}
\description{
Sets the current api token. Called after authentication succeeds
}
