% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{.parse_contents}
\alias{.parse_contents}
\title{.parse_contents}
\usage{
.parse_contents(res)
}
\arguments{
\item{res}{the response object from the httr::GET method}
}
\value{
the parsed contents object
}
\description{
parses the contents of a fetch request. Will raise errors if response is above 400. Will also
clean any hyperlinked bdbids. Other clean code for specific requests can get added to this method.
}
