% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{fetch_request}
\alias{fetch_request}
\title{fetch_request}
\usage{
fetch_request(endpoint_url, query_params = NULL)
}
\arguments{
\item{with_caching}{}
}
\value{
parsed contents from the fetch request
}
\description{
fetch's from one of bpl's read-only list views using GET and
returns 1:n records as a dataframe or raises an http_error
}
\details{
TODO create a paginating endpoint similar to this for large responses
}
