% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{polling_request}
\alias{polling_request}
\title{polling_request}
\usage{
polling_request(endpoint_url, task_id, tries = 10,
  polling_interval = 3)
}
\arguments{
\item{endpoint_url}{the endpoint to fetch results}

\item{tries}{the number of GET requests to send before giving up}

\item{polling_interval}{the timeout interval between requests}
}
\value{
parsed result or failure
}
\description{
Polling endpoint for urls to fetch completed work from bema
}
