% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{post_request}
\alias{post_request}
\title{post_request}
\usage{
post_request(endpoint_url, payload = list())
}
\arguments{
\item{payload}{}
}
\description{
A post request route used for sending a data request to bema endpoints. Takes
a payload argument with a list for various settings documented in the service
}
